+++ src/neurobrix/kernels/utils/triton_driver_helper.py
@@ -1,8 +1,0 @@
-try:
-    from neurobrix.kernels.runtime import torch_device_fn
-
-    get_device_properties = torch_device_fn.get_device_properties
-except AttributeError:
-    import triton
-
-    get_device_properties = triton.runtime.driver.active.utils.get_device_properties